Updated to version 1.3.2 to fix save issue
git-svn-id: https://plugins.svn.wordpress.org/add-logo-to-admin/trunk@98743 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ add-logo.php
@@ -4,7 +4,7 @@ Plugin Name: Add Logo to Admin
 Plugin URI: http://bavotasan.com/tidbits/add-your-logo-to-the-wordpress-admin-and-login-page/
 Description: Adds a custom logo to your site's Admin header and your login page.
 Author: c.bavota
-Version: 1.3.1
+Version: 1.3.2
 Author URI: http://bavotasan.com
 */
 
@@ -14,6 +14,7 @@ add_action('admin_menu', 'add_logo_init');
 
 //add page to admin 
 function add_logo_init() {
+if(stristr($_SERVER['REQUEST_URI'],'?page=add-logo-to-admin/add-logo.php')) {
 	if(stristr($_SERVER['REQUEST_URI'],'&delete-logo')) {
 		$directory = dirname(__FILE__) . "/images/";
 		unlink($directory . $_REQUEST['delete-logo']);
@@ -59,7 +60,7 @@ function add_logo_init() {
 		die;
 	}
 	
-	
+}	
   add_options_page('Add Logo to Admin', 'Add Logo to Admin', 10, __FILE__, 'my_plugin_options');
 }
